Updated dialog UI and added Tor download link
This commit updates the dialog UI by changing the icon color to tertiary. It also adds a "Download Orbot" button to the Tor disabled dialog, which links to the Orbot app on the Google Play Store.

Additionally, the version name has been updated to 4.0.0 and the version code to 30011.

The commit also includes a fix to ensure the arrow in the MainMediaFragment is shown/hidden correctly when the ContentPickerFragment is dismissed.